Endret navn på tårn til linjeforeningene + addet forskjellige rewards for å drepe forskjellige typer creeps

+++ core/src/no/fusiontd/components/Durability.java
@@ -3,9 +3,9 @@ package no.fusiontd.components;
 import no.fusiontd.CloneableComponent;
 
 public class Durability implements CloneableComponent<Durability> {
-    public float life = 1;
+    public int life = 1;
 
-    public Durability(float maxLife) {
+    public Durability(int maxLife) {
         this.life = maxLife;
     }
